refactor: changed api/sql name to api/resources and fixed imports
