+++ api/resources/light_curve/light_curve.py
