+++ api/resources/light_curve/models.py
