+++ api/resources/magstats/magstats.py
