+++ api/resources/magstats/models.py
